worker/requirements.txt
httpx==0.28.1
psycopg[binary]==3.2.10
python-dotenv==1.0.1
